requirements.txt
flask
pillow
torch==2.0.1+cu118
torchvision==0.15.2+cu118
diffusers==0.21.4
transformers
accelerate
xformers==0.0.20